:truck: Change module structure
