+++ src/lib.rs
@@ -6,12 +6,15 @@
 #[macro_use]
 pub(crate) mod test_utils;
 
-pub(crate) mod common;
+mod gf2_linalg;
+mod utils;
+mod validate;
+
+pub mod common;
 pub mod flow;
 pub(crate) mod gf2_linalg;
 pub mod gflow;
 pub mod pflow;
-pub(crate) mod validate;
 
 use pyo3::prelude::*;
